move ThemeManager into Themes/ folder
