+++ app/renderer/src/plugins/PluginLoader.ts
@@ -13,7 +13,7 @@ import Bridge from './PluginEnv'
 import EventBus from '../EventBus'
 import { PluginSnippets } from '../../windows/Snippets'
 import { UI_DATA, BridgeCore } from '../bridgeCore/main'
-import ThemeManager from '../editor/ThemeManager'
+import ThemeManager from '../editor/Themes/ThemeManager'
 import unzipper from 'unzipper'
 import safeEval from 'safe-eval'
 import ComponentRegistry from './CustomComponents'
