+++ app/renderer/windows/Settings.ts
@@ -12,7 +12,7 @@ import AddSnippetWindow from './AddSnippet'
 import Snippets from './Snippets'
 import { ipcRenderer, remote } from 'electron'
 import ConfirmWindow from '../src/commonWindows/Confirm'
-import ThemeManager from '../src/editor/ThemeManager'
+import ThemeManager from '../src/editor/Themes/ThemeManager'
 import ProjectConfig from '../src/Project/Config'
 import { uuid } from '../src/Utilities/useAttr'
 import FontList from 'font-list'
